formulaire edit user

+++ src/Controller/Admin/AdminUserController.php
@@ -5,8 +5,10 @@ namespace App\Controller\Admin;
 use App\Entity\User;
 use App\Entity\Product;
 use App\Entity\Supplier;
+use App\Form\UserType;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -23,14 +25,31 @@ class AdminUserController extends AbstractController
     }
 
     #[Route('/edit/{id}', name: 'edit')]
-    public function edit(): Response
+    public function edit(User $user, Request $request, EntityManagerInterface $em): Response
     {
-        return $this->render('admin/user/edit.html.twig');
+        $this->denyAccessUnlessGranted('USER_EDIT', $user);
+
+        $userForm = $this->createForm(UserType::class, $user);
+
+        $userForm->handleRequest($request);
+
+        if($userForm->isSubmitted() && $userForm->isValid()){
+            $em->persist($user);
+            $em->flush();
+
+            return $this->redirectToRoute('admin_user_index');
+        }
+
+        return $this->render('admin/user/edit.html.twig', [
+            'userForm' => $userForm->createView()
+        ]);
     }
 
     #[Route('/suppression/{id}', name: 'delete')]
     public function delete(User $user, EntityManagerInterface $em): Response
     {
+        $this->denyAccessUnlessGranted('USER_DELETE', $user);
+
         $em->remove($user);
         $em->flush();
